일부 에러 발생 수정 및 게시글 type 추가. jpa cusotom method 작성위한 상속단계 추가

+++ HarmonyLink-SpringBoot/src/main/java/com/dongyang/HarmonyLink/MVC/Service/RegisterService.java
@@ -4,8 +4,10 @@ import com.dongyang.HarmonyLink.MVC.domain.User.DTO.UserDTO;
 import com.dongyang.HarmonyLink.MVC.domain.User.Entity.UserEntity;
 import com.dongyang.HarmonyLink.MVC.Repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @Slf4j
@@ -18,19 +20,18 @@ public class RegisterService {
     }
 
     public UserDTO getInfo(String id) {
-        // 추후 Optional 부분 공부 후에 orElse 대신, Optional<> Wrapper Class로 수행하기
-        UserEntity entity = userRepository.findById(id).orElse(null);
+        UserEntity entity = userRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "존재하지 않는 회원입니다."));
 
         return UserDTO.toDTO(entity);
     }
 
     @Transactional
     public UserDTO register(UserDTO dto) {
-        UserEntity userEntity = UserEntity.toEntity(dto);
 
         // 해당 사용자가 이미 등록된 경우
-        if(userRepository.checkAlreadyRegister(dto.getId(), dto.getEmail()).orElse(null) != null) {
-            throw new RuntimeException("회원가입이 이미 완료된 회원입니다."); // 등록을 수행하지 않음
+        if(userRepository.checkAlreadyRegister(dto.getId(), dto.getEmail()).isPresent()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"회원가입이 이미 완료된 회원입니다."); // 등록을 수행하지 않음
         }
 
         // DTO to Entity + save
@@ -48,10 +49,8 @@ public class RegisterService {
 
     @Transactional
     public UserDTO patchUser(UserDTO dto) {
-        UserEntity userEntity = UserEntity.toEntity(dto);
-
-        // 예외처리 부문
-
+        if(userRepository.findByUserKey(dto.getUserKey()).isEmpty())
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "존재하지 않는 회원입니다.");
 
         UserEntity result = userRepository.save(UserEntity.toEntity(dto));
 
@@ -60,18 +59,9 @@ public class RegisterService {
 
     @Transactional
     public void deleteUser(UserDTO dto) {
+        if(userRepository.findByUserKey(dto.getUserKey()).isEmpty())
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "존재하지 않는 회원입니다.");
+
         userRepository.delete(UserEntity.toEntity(dto));
     }
-
-//    public UserDTO patch(UserDTO dto) {
-//        UserEntity target = registerRepository.findById(dto.getId())
-//                .orElseThrow(() -> new IllegalArgumentException("회원 정보 수정 실패. 존재하지 않는 사용자입니다."));
-//
-//        // dto 매핑 관련 영상 시청 후 수행하기.
-//        // DTO, Entity에 mapper method 작성하는 건 올바르지 못함.
-//
-//
-//        return UserDTO.toDTO();
-//
-//    }
 }
